dywinlilab.ipynb, IagoBot_ValueNet.ipynb
dywinlilab.ipynb [In Progress] - The Dynamic Win-Likelihood Labeler reads in successive Othello game files and dynamically reconstructs an 8x8 representation of the board state for each move according to Othello rules. If a duplicated board state is found then the likelihood of winning for that state is calculated as (#of wins from that state)/(total # of such states).

IagoBot_ValueNet.ipynb [In Progress] - The Value Network for IagoBot is a convolutional neural network that outputs the likelihood of winning for a given board position.